Created files and fixed orthography
Now I'm able to understand and manage different kinds of objects
whose are the milestones of the C language.
I'm also trying a new kind of workflow to be able to manage many
files at the same time. To keep reading and thinking about the
possibilities on adapting these old programs made me feel more
relaxed about the general structure of this language.
Since this is only a bunch of study files I don't think they can
grow up on it's own to become a good project. But the idea was always
to integrate the speed and flexibility of C with the python structure
and high-level facilities.