Learning to build a real Project
This, to me, is the last barrier to build a project. Never played with
compiled languages, and never build a project in C, so it won't be that
easy to understand whats the similarities and differences between what I
know and what I need to know to make it work.