Chapter Four Finished.
Now I'm rushing at learning C language. I almost can do anything that
I'm used to in python, and that wonderful. With the ability to handle
data, write functions, compile code within different source files, I'm
almost there. The next step is to use all the power of C to see what it
is able to. Obviously it's one of the most powerful languages, but also
one of the harder. To keep track of a study like this one is also hard,
and to choose what is important to me or not is what keeps me going. I
know at least what I could use. Oh yeah, half of the book is gone. Now
I intend to to advance to the second half. Here we go.