Added comments, made a few changes
Now that I'm better knowing what I'm actually doing at those lines I'm
introducing some comments and some changes to make it more...
interesting to came back later. All those files are a way to learn, and
to reproduce and test with them is a way to make sure that I've
learned.