Did a few changes at NumberOne
I'm reading a book translated to portuguese, so most of the times I need
to translate what's inside the functions. On NumberOne I needed to
translate "tam" that stands for "tamanho" to "size".
On NumberFive I've created some new files and updated the Session.vim to
save the file on my default screen.