I don't care about what I did
I did a lot of things. From adding comments to older files, to creating
from scratch a lot of new files. This is not exactly a project of
production, but a lot of files that I'm writing mainly to be easy to
understand and re-visit later when I'm desperate to know "Holy, how I
did it yesterday?".