Added stuff to the structures chapter and more
Since I'm slowly learning, and I have to stop, sometimes I forget about
the lessons, and it becomes a pain in my a$$. It's hard, but I'll reach
the end of the book, and probably add whatever I find at the other two
books. Any information is good information about C language. It's really
hard to find a good source to learn it alone.